renamed aws resouces to follow naming conventions
didn't know what to name the gcp resource, so left them as is for now
